Add config, wheel parser and libraries to crosswrench
Mega commit.
It includes:
Config reader that reads settings from the python
interpreter.
Wheel parser that takes a whole WHEEL file as data
and makes the variables usable from C++.
Skeleton for the main execute runner.
The libraries libzippp, pstreams and pystring.
CMake Find modules so that pstreams and pystring
can be found when linking to them as external
libraries.