Add record class constructor and it deps to crosswrench
Add the initial part for record handling class.
Initially just the constructor.
record uses the cvs2 header library, so this commit
adds csv2 to libs including changes to LICENSE files.
record needs .dist-info dirname so functions to
get this are also added in this commit.
Since std::filesystem is used in the dirname
functions, require usage of C++17.
hashlib2botan translator of hashlib names are also
provided as record needs to verify these.
hashlib2botan uses botan so link it in.
clang-format changes so that botan, csv2 and pystring
headers get sorted correctly.