Start real work on execute
Start real work on execute.
Make the execute function verify
the wheel filename, open the
wheel passed on command line and
verify it.
Remove all other cruft in execute.
Change the signature on functions
and methods that take ZipArchive
as argument since it must be used
with & (no copy constructor).