Fix bugs in record handling, due to API missunderstandings
Fix bugs in record handling, due to API missunderstandings.
The second argument to hasEntry excludeDirectories does not
exclude directories but ignores the directory part of the
name requested. So set this to false (the default) to make
it work correctly.
readEntry does not return the number of bytes read but
a any of the LIBZIPPP_* constants. End if not LIBZIPPP_OK.
readEntry returns LIBZIPPP_ERROR_FREAD_FAILURE when reading
of size zero. Just not executing readExecute is fine in
this case since hasher->final() will return the right
value even if update has never been called.
In order to fix these bugs output from each failure was
needed in record::verify(libzippp::ZipArchive&).