Initial install step in crosswrench
Initial install step in crosswrench.
Add a spread class so named after
the spread step in PEP427.
This class encapulates the spreading
of files step. If one method per
directory is needed remains to be
seen and using an std::map with
std::functions in it does not seem
to work as expected.
Currently only an initial
implementation of installroot is
provided, the other are stubs.
No error handling is done in this
function.
Added isrecordfilenames that checks
if a filename is one that should
not be in RECORD.