Add support for adding records and writing the RECORD file to the record class
Add support for adding records and writing the RECORD file to the record class.
Since both installroot in spread and write in record uses the location where
the root of the wheel is installed, a function is added that generates it
to avoid code duplication.
fix bug that made the files under .info dir not being installed.
clang-formating and some other changes.