Generalize installation methods
Generalize installation methods.
Add installfile method to spread
this is a general install method
based on the former work in
installroot.
Add installpath that gives a path
based on the python config.
installpath is the last half of
rootinstallpath to us it in
rootinstallpath.
Remove all the stub install
methods since they are useless.
Remove installroot.