Speed up bytecompilation step
Speed up bytecompilation step.
The byte compilation step
started a python interpreter
for every .py file it was to
compile.
Instead start one interpreter
and give it a list of files
via pipe.
In order to make this work
the get_cmd_output function
had to be changed to also
take pipe input. This input
is only piped to the process
if it contains something.
The places where
get_cmd_output is used had to
be changed according to the
new signature, these places
use an empty pipe argument.