Install files from .data directory in the right place
Install files from .data directory in the right place.
Files in {distribution}-{version}.data/{type} are not
to be placed relative to the root but relative to
{distribution}-{version}.data/{type}.
The signature for installfiles changes due to this
so it takes the last part of the part as an argument
instead of hard coding it to entry.getName().