Fix installation of headers and verify .data content
Fix installation of headers and verify .data content.
The direcory for headers does not have the same name
as the key to look it up. Add a map that maps dirs to
a config key. The config constructor needs to be
implemented due to this instead of using default.
Add dotdatainstallpath that returns a path out of the
directory key.
Add a function that verifies that only mapable dirs
exists in the .data directory. Use this function in
the initial verification of the wheel.