Installer rewrite, generate install locations only and use that in installfile
Installer rewrite, generate install locations only and use that in installfile.
Instead of having install methods that generate install prefixes then
calls installfile with this prefix, make them methods that return the whole
install path and make installfile take the installpath as an argument.
Rename the methods accordingly.
The idea behind this is that control functions can later be written to check
that all entries in the wheel can be installed before actually touching
the filesystem, install location generation must be independent from file
writing for this to work.
Since the path generating method have been renamed to the same names as
some functions in functions.cpp, change the name of the functions and change
lines where they are called.
Use a function that takes a ZipEntry to check if a ZipEntry is a script
instead of using a bool to installfile.