Fix paths in installed RECORD
Fix paths in installed RECORD.
Before this commit the paths
in the installed RECORD file
where the same as in the wheel
file.
The standard for the installed
RECORD file specifies that the
paths must either be absolute
or relative to the directory
where the .dist-info directry
resides.
The solution used is to make
all paths relative.