Add support for writing the INSTALLER file
Add support for writing the INSTALLER file.
In order to do this a new overload to
method installfile is added. This one does
take memory data instead of a zip entry and
installs the data into the file location
in filepath. It is based on the code in
installfile that takes a zip entry.