Fix use of external pstreams
Fix use of external pstreams.
The pstreams version test had
a ) in it that caused it to
always fail.
Also add the main function
since cmake docs claim it is
required.
Use INTERFACE instead of
IMPORTED since it is a header
only library and no installed
location is needed.
This removes the warning
about installed location and
and fixes errors like
"target pattern contains no %"
from gmake.