Use the strongest hash in installed RECORD and fix SHA-3 handling
Use the strongest hash in installed RECORD and fix SHA-3 handling.
There was a typo in the name used to create sha3 hashes using
botan, fix it, SHA3->SHA-3.
Instead of using SHA256 in the installed RECORD file, use
the stongest one available in algotrithms_guaranteed.
strongest_algorithm_hashlib gives the hashlib name of the algorithm.
strongest_algorithm_botan gives the botan name of the same.
In order to do this algorithms_guaranteed need to be fetched
from python. This is done in get_algos in the config class.
While in the config class a unused variable errors was removed.
The algos from python are then transformed into a vector of
algorithm names.
The function strvec_contains that checks if a string is a stored
in a vector of strings was created to make finding the strongest
algorithm easier.
The print_guaranteed method of hashlib2botan class was created for
debugging.