Add elf identification of files
Add elf identification of files.
Elf files like shared libraries
and executables are usually
installed with exec rights set.
Wheels are zip files and there
are extra fields that can include
this information but this is
not guranteed to be set.
After having downloaded and
checked wheels that include .so
files and found that they did not
include these extra fields it
was decided to go with this
solution instead.
Elf files of the types shared
object and executable are now
identified on the fly when
unpacked.