Add support creating entrypoint console scripts
Add support creating entrypoint console scripts.
This is done to support the entry points
specification.
Three functions are used for this.
createscript that just concatinates text to create
the script content based on the information on
the right side of the equal sign.
getentrypointscripts that takes an entrypoints
file and create a map with the script name and
content of the script. It uses createscript to
create the content part.
installentrypointconsolescripts in spread class
that checks if entry_points.txt exists and if so
uses getentryointscripts get the names and scripts
and if any scripts are generated installes the
scripts in their respective names and makes them
executable.
Since filesystem exec permissions are now set in
two places, make it a function called setexecperms
since this function call is quite big.