Reduce calls to the python interpreter, change get_cmd_output signature
When crosswrench calls the python interpreter 7 times when it
extracts install information from python. This is because it
calls the python interpreter for every variable it needs info
from.
Instead call the interpreter to output all variables at once
leading to just one call.
In order to get multiline output get_cmd_output had to be
changed to enable this. It now takes a vector of strings
instead of a string as output.
All users of this function have been changed accordingly.