Make reading from the python process more robust
Make reading from the python process more robust.
get_cmd_output did not read from stderr so python
interpreters that output the version string to
stderr (jython) crashed crosswrench. So make it
read from both.
Make output include an empry string if the
command does not print anything, this is needed
since the code in crosswrench assumes that the
output argument to get_cmd_output is not an
empty vector after executation of get_cmd_output.