Check for access permissions and fix zero size entries
Add function that checks for access permissions and
use it on all files before installing.
Don't use readEntry on zero sized file entries since
readEntry returns LIBZIPPP_ERROR_FREAD_FAILURE when
used, not writing anything is safe since nothing needs
to be written to create the right file and get the
correct hash.
Add an installpath overload that takes a string that
is the filename in the .distinfo directory and returns
the full path to it. For usage with the access
permission check function.